feedbacks

+++ assets/js/BulletinItem.js
@@ -10,6 +10,8 @@
 import Page from 'teamelf/layout/Page';
 const { Row, Col, Button, Input, Checkbox, TreeSelect, Icon } = antd;
 import BulletinProcess from 'teamelf/bulletin/BulletinProcess';
+import BulletinPreview from 'teamelf/bulletin/BulletinPreview';
+import BulletinFeedback from 'teamelf/bulletin/BulletinFeedback';
 
 export default class extends Page {
   constructor (props) {
@@ -20,26 +22,56 @@ export default class extends Page {
       publishing: false,
       deleting: false,
       changed: false,
-      bulletin: null
+      refreshing: false,
+      autoRefresh: false,
+      mentionList: [],
+      bulletin: null,
+      feedbacks: []
     };
+    this.fetchMentionList();
     this.fetchBulletin();
+  }
+  async fetchMentionList () {
+    let roles = (await axios.get('role')).data.map(o => ({
+      label: o.name,
+      value: 'r_' + o.slug,
+      key: 'r_' + o.slug,
+      children: []
+    }));
+    for (let member of (await axios.get('member')).data) {
+      let role = roles.find(o => o.value === 'r_' + member.role.slug);
+      role.children.push({
+        label: member.name,
+        value: 'm_' + member.username,
+        key: 'm_' + member.username
+      });
+    }
+    this.setState({mentionList: roles});
+    this.fetchFeedback();
   }
   fetchBulletin () {
     const id = this.props.match.params.id;
     return axios.get('bulletin/' + id).then(r => {
       this.setState({bulletin: r.data});
-      this.autoSave();
+      if (r.data.isDraft) {
+        this.autoSave();
+      } else {
+        this.setState({autoRefresh: true});
+        this.autoRefreshFeedback();
+      }
     });
   }
   save () {
     const data = {
       title: this.state.bulletin.title || '',
       content: this.state.bulletin.content || '',
+      receivers: this.state.bulletin.receivers || []
     };
     const id = this.props.match.params.id;
     this.setState({saving: true});
     return axios.put('bulletin/' + id, data).then(r => {
       this.setState({saving: false, changed: false});
+      this.fetchBulletin();
     }).catch(e => {
       this.setState({saving: false});
     });
@@ -49,20 +81,25 @@ export default class extends Page {
     await this.save();
     this.setState({publishing: true});
     return axios.put(`bulletin/${id}/publish`).then(r => {
-      this.setState({publishing: false});
-      this.fetchBulletin();
+      window.location.reload();
     }).catch(e => {
       this.setState({publishing: false});
     });
   }
   del () {
-    const id = this.props.match.params.id;
-    this.setState({deleting: false});
-    return axios.delete('bulletin/' + id).then(r => {
-      this.setState({deleting: false});
-      window.location.href = '/bulletin';
-    }).catch(e => {
-      this.setState({deleting: false});
+    antd.Modal.confirm({
+      title: '不可恢复',
+      content: '确定要删除么？该操作可能无法恢复',
+      onOk: () => {
+        const id = this.props.match.params.id;
+        this.setState({deleting: true});
+        return axios.delete('bulletin/' + id).then(r => {
+          this.setState({deleting: false});
+          window.location.href = '/bulletin';
+        }).catch(e => {
+          this.setState({deleting: false});
+        });
+      }
     });
   }
   async autoSave () {
@@ -70,6 +107,22 @@ export default class extends Page {
       await this.save();
     }
     setTimeout(this.autoSave.bind(this), 60000);
+  }
+  fetchFeedback () {
+    const id = this.props.match.params.id;
+    this.setState({refreshing: true});
+    return axios.get(`bulletin/${id}/feedback`).then(r => {
+      this.setState({feedbacks: r.data});
+      this.setState({refreshing: false});
+    }).catch(e => {
+      this.setState({refreshing: false});
+    });
+  }
+  async autoRefreshFeedback () {
+    if (this.state.autoRefresh) {
+      await this.fetchFeedback();
+    }
+    setTimeout(this.autoRefreshFeedback.bind(this), 60000);
   }
   title () {
     if (this.state.bulletin) {
@@ -91,6 +144,7 @@ export default class extends Page {
             </div>
           </Col>
           <Col xs={24} md={12}>
+            {this.state.bulletin.isDraft &&
             <div style={{marginBottom: 16}}>
               <Checkbox
                 checked={this.state.autoSave}
@@ -100,84 +154,110 @@ export default class extends Page {
                 : '自动存草稿'
               }</Checkbox>
             </div>
+            }
+            {!this.state.bulletin.isDraft && [
+              <div style={{marginBottom: 16}}>
+                <Checkbox
+                  checked={this.state.autoRefresh}
+                  onChange={e => this.setState({autoRefresh: e.target.checked})}
+                >自动刷新反馈</Checkbox>
+              </div>,
+              <div style={{marginBottom: 16}}>
+                <Button
+                  type="primary"
+                  icon="reload"
+                  onClick={this.fetchFeedback.bind(this)}
+                  loading={this.state.refreshing}
+                >刷新反馈</Button>
+              </div>
+            ]}
             {this.state.bulletin.isDraft &&
-              <Row type="flex" justify="start" gutter={16}>
-                <Col>
-                  <Button
-                    type="primary"
-                    onClick={this.save.bind(this)}
-                    loading={this.state.saving}
-                  >保存草稿</Button>
-                </Col>
-                <Col>
-                  <Button
-                    type="primary"
-                    onClick={this.publish.bind(this)}
-                    loading={this.state.publishing}
-                  >发布公告</Button>
-                </Col>
-                <Col>
-                  <Button
-                    type="danger"
-                    onClick={this.del.bind(this)}
-                    loading={this.state.deleting}
-                  >舍弃</Button>
-                </Col>
-              </Row>
+            <Row type="flex" justify="start" gutter={16}>
+              <Col>
+                <Button
+                  type="primary"
+                  onClick={this.save.bind(this)}
+                  loading={this.state.saving}
+                >保存草稿</Button>
+              </Col>
+              <Col>
+                <Button
+                  type="primary"
+                  onClick={this.publish.bind(this)}
+                  loading={this.state.publishing}
+                >发布公告</Button>
+              </Col>
+              <Col>
+                <Button
+                  type="danger"
+                  onClick={this.del.bind(this)}
+                  loading={this.state.deleting}
+                >舍弃</Button>
+              </Col>
+            </Row>
             }
           </Col>
         </Row>
       );
     }
   }
-  handleBulletinChange (key, e) {
+  handleBulletinChange (key, value) {
     const bulletin = this.state.bulletin;
-    bulletin[key] = e.target.value;
+    bulletin[key] = value;
     this.setState({bulletin, changed: true});
   }
-  renderView () {
+  renderEditor () {
     return (
-      <Row type="flex" gutter={16}>
-        <Col xs={24} md={12}>
-          <div style={{marginBottom: 16}}>
-            <Input
-              size="large"
-              value={this.state.bulletin.title}
-              onChange={this.handleBulletinChange.bind(this, 'title')}
-            />
-          </div>
-          <div style={{marginBottom: 16}}>
-            <TreeSelect
-              size="large" style={{width: '100%'}}
-              treeCheckable={true}
-              showCheckedStrategy={TreeSelect.SHOW_PARENT}
-              searchPlaceholder="选择要通知的人"
-              treeData={[]}
-              value={[]}
-              onChange={e => console.log(e)}
-            />
-          </div>
-          <div style={{marginBottom: 16}}>
-            <Input.TextArea
-              size="large"
-              autosize={{minRows: 10, maxRows: 30}}
-              value={this.state.bulletin.content}
-              onChange={this.handleBulletinChange.bind(this, 'content')}
-            />
-          </div>
-        </Col>
-        <Col xs={24} md={12}>
-          <div>{this.state.bulletin.title}</div>
-          <div>{this.state.bulletin.content}</div>
-        </Col>
-      </Row>
+      <div>
+        <div style={{marginBottom: 16}}>
+          <Input
+            size="large"
+            value={this.state.bulletin.title}
+            onChange={e => this.handleBulletinChange('title', e.target.value)}
+          />
+        </div>
+        <div style={{marginBottom: 16}}>
+          <TreeSelect
+            size="large" style={{width: '100%'}}
+            treeCheckable
+            showCheckedStrategy={TreeSelect.SHOW_PARENT}
+            searchPlaceholder="选择要通知的人"
+            treeData={this.state.mentionList}
+            value={this.state.bulletin.receivers}
+            onChange={e => this.handleBulletinChange('receivers', e)}
+            allowClear
+          />
+        </div>
+        <div style={{marginBottom: 16}}>
+          <Input.TextArea
+            size="large"
+            autosize={{minRows: 10, maxRows: 30}}
+            value={this.state.bulletin.content}
+            onChange={e => this.handleBulletinChange('content', e.target.value)}
+          />
+        </div>
+      </div>
     );
   }
   view () {
     if (this.state.bulletin) {
       return [
         <BulletinProcess isDraft={this.state.bulletin.isDraft}/>,
-        this.renderView()
+        <Row type="flex" gutter={16}>
+          {this.state.bulletin.isDraft &&
+          <Col xs={24} md={12}>
+            {this.renderEditor()}
+          </Col>
+          }
+          <Col xs={24} md={12}>
+            <BulletinPreview {...this.state.bulletin}/>
+          </Col>
+          {!this.state.bulletin.isDraft &&
+          <Col xs={24} md={12}>
+            <BulletinFeedback feedbacks={this.state.feedbacks}/>
+          </Col>
+          }
+        </Row>
       ];
     }
   }
